+++ src/app/dashboard/configuration/assets-category/edit/[id]/page.tsx
@@ -1,11 +1,10 @@
-import { EditAssetsCategory } from '@/modules/configuration/assets_category';
-import { funGetOneAssetsCategory } from '@/modules/configuration/assets_category/fun/assets_category_get_one';
+import { EditAssetsCategory, funGetOneAssetsCategory } from '@/modules/assets';
 import React from 'react';
 
-export default async function Page({params}: {params: {id: string}}) {
-  const data = await funGetOneAssetsCategory(params.id)
-  if(!data) return <></>
+export default async function Page({ params }: { params: { id: string } }) {
+  const data = await funGetOneAssetsCategory({ id: params.id });
+  if (!data) return <></>
   return (
-   <EditAssetsCategory data={data}/>
+    <EditAssetsCategory data={data} />
   );
 }
