+++ src/app/dashboard/configuration/assets-category/page.tsx
@@ -1,11 +1,7 @@
-import {
-  TableAssetsCategory,
-  funGetAllAssetsCategory,
-} from "@/modules/configuration/assets_category";
-import { DefaultProps } from "@mantine/core";
+import { TableAssetsCategory, funGetAllAssetsCategory } from "@/modules/assets";
 import React from "react";
 
 export default async function Page() {
-  const data = await funGetAllAssetsCategory(1);
-  return <TableAssetsCategory data={data}  />;
+  const data = await funGetAllAssetsCategory({ page: 1 });
+  return <TableAssetsCategory data={data} />;
 }
