--- src/modules/assets/modules/index.ts
@@ -1,19 +1,0 @@
-import DetailAssets from "./detail_assets/detail_assets";
-import DetailAssetsHistory from "./detail_assets_history/detail_assets_history";
-import CreateAssets from "./create_assets/create_assets";
-import EditAssets from "./edit_assets/edit_assets";
-import ModalEditAssets from "./edit_assets/modal_edit_assets";
-import ModalViewBarcode from "./detail_assets/view_barcode/modal_view_barcode";
-import ViewBarcode from "./detail_assets/view_barcode/view_barcode";
-import HistoryScan from "./scan/history_scan";
-import ScanBarcode from "./scan/scan_barcode";
-
-export {DetailAssets}
-export {DetailAssetsHistory}
-export {CreateAssets}
-export {EditAssets}
-export {ModalEditAssets}
-export {ModalViewBarcode}
-export {ViewBarcode}
-export {HistoryScan}
-export {ScanBarcode}
