--- src/modules/assets/modules/val/idModalAssets.ts
@@ -1,4 +1,0 @@
-"use client";
-import { atom } from "jotai";
-
-export const isModalAssetsView = atom(false);
