--- src/modules/assets/modules/val/isModalAssetsEdit.ts
@@ -1,5 +1,0 @@
-"use client";
-
-import { atom } from "jotai";
-
-export const isModalAssetsEdit = atom(false);
