--- src/modules/assets/modules/val/isModalBarcode.ts
@@ -1,5 +1,0 @@
-"use client";
-
-import { atom } from "jotai";
-
-export const isModalBarcode = atom(false);
